+++ job/common/src/test/java/alluxio/job/workflow/composite/CompositeConfigTest.java
@@ -9,7 +9,7 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.job.composite;
+package alluxio.job.workflow.composite;
 
 import static junit.framework.TestCase.assertNotNull;
 import static org.junit.Assert.assertEquals;
