+++ job/server/src/main/java/alluxio/job/workflow/composite/CompositeTask.java
@@ -9,7 +9,7 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.job.composite;
+package alluxio.job.workflow.composite;
 
 import alluxio.job.JobConfig;
