+++ job/server/src/main/java/alluxio/master/job/plan/PlanCoordinator.java
@@ -9,7 +9,7 @@
  * See the NOTICE file distributed with this work for information regarding copyright ownership.
  */
 
-package alluxio.master.plan;
+package alluxio.master.job.plan;
 
 import alluxio.exception.JobDoesNotExistException;
 import alluxio.job.JobConfig;
