+++ job/server/src/main/resources/META-INF/services/alluxio.job.JobDefinition
@@ -1,4 +1,4 @@
-alluxio.job.composite.CompositeDefinition
+alluxio.job.workflow.composite.CompositeDefinition
 alluxio.job.plan.load.LoadDefinition
 alluxio.job.plan.migrate.MigrateDefinition
 alluxio.job.plan.persist.PersistDefinition
